// src/domains/service-clients/payment/proto/generated/payment_service.ts
export interface CreatePaymentRequest {
  provider: Provider;
  userId: string;
  orderId: string;
  /** For ensuring idempotency */
  idempotencyKey: string;
  successUrl?: string | undefined;
  cancelUrl?: string | undefined;
}
// ...
export const CreatePaymentRequest: MessageFns<CreatePaymentRequest> = {
  encode(
    message: CreatePaymentRequest,
    writer: BinaryWriter = new BinaryWriter()
  ): BinaryWriter {
    if (message.provider !== 0) {
      writer.uint32(8).int32(message.provider);
    }
    if (message.userId !== '') {
      writer.uint32(18).string(message.userId);
    }
    if (message.orderId !== '') {
      writer.uint32(26).string(message.orderId);
    }
    if (message.idempotencyKey !== '') {
      writer.uint32(42).string(message.idempotencyKey);
    }
    if (message.successUrl !== undefined) {
      writer.uint32(58).string(message.successUrl);
    }
    if (message.cancelUrl !== undefined) {
      writer.uint32(66).string(message.cancelUrl);
    }
    return writer;
  },

  decode(
    input: BinaryReader | Uint8Array,
    length?: number
  ): CreatePaymentRequest {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreatePaymentRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.provider = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.userId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.orderId = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.idempotencyKey = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.successUrl = reader.string();
          continue;
        }
        case 8: {
          if (tag !== 66) {
            break;
          }

          message.cancelUrl = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreatePaymentRequest {
    return {
      provider: isSet(object.provider) ? providerFromJSON(object.provider) : 0,
      userId: isSet(object.userId) ? globalThis.String(object.userId) : '',
      orderId: isSet(object.orderId) ? globalThis.String(object.orderId) : '',
      idempotencyKey: isSet(object.idempotencyKey)
        ? globalThis.String(object.idempotencyKey)
        : '',
      successUrl: isSet(object.successUrl)
        ? globalThis.String(object.successUrl)
        : undefined,
      cancelUrl: isSet(object.cancelUrl)
        ? globalThis.String(object.cancelUrl)
        : undefined,
    };
  },

  toJSON(message: CreatePaymentRequest): unknown {
    const obj: any = {};
    if (message.provider !== 0) {
      obj.provider = providerToJSON(message.provider);
    }
    if (message.userId !== '') {
      obj.userId = message.userId;
    }
    if (message.orderId !== '') {
      obj.orderId = message.orderId;
    }
    if (message.idempotencyKey !== '') {
      obj.idempotencyKey = message.idempotencyKey;
    }
    if (message.successUrl !== undefined) {
      obj.successUrl = message.successUrl;
    }
    if (message.cancelUrl !== undefined) {
      obj.cancelUrl = message.cancelUrl;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CreatePaymentRequest>, I>>(
    base?: I
  ): CreatePaymentRequest {
    return CreatePaymentRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreatePaymentRequest>, I>>(
    object: I
  ): CreatePaymentRequest {
    const message = createBaseCreatePaymentRequest();
    message.provider = object.provider ?? 0;
    message.userId = object.userId ?? '';
    message.orderId = object.orderId ?? '';
    message.idempotencyKey = object.idempotencyKey ?? '';
    message.successUrl = object.successUrl ?? undefined;
    message.cancelUrl = object.cancelUrl ?? undefined;
    return message;
  },
};
// ...
type Builtin =
  | Date
  | Function
  | Uint8Array
  | string
  | number
  | boolean
  | undefined;

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P>>]: never;
    };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}