// src/domains/service-clients/payment/proto/generated/payment/common.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.11.1
//   protoc               v5.29.6
// source: payment/common.proto

/* eslint-disable */

export const protobufPackage = "payment.common";

export enum Provider {
  PROVIDER_UNSPECIFIED = 0,
  STRIPE = 1,
  RAZORPAY = 2,
  PAYPAL = 3,
  UNRECOGNIZED = -1,
}

export function providerFromJSON(object: any): Provider {
  switch (object) {
    case 0:
    case "PROVIDER_UNSPECIFIED":
      return Provider.PROVIDER_UNSPECIFIED;
    case 1:
    case "STRIPE":
      return Provider.STRIPE;
    case 2:
    case "RAZORPAY":
      return Provider.RAZORPAY;
    case 3:
    case "PAYPAL":
      return Provider.PAYPAL;
    case -1:
    case "UNRECOGNIZED":
    default:
      return Provider.UNRECOGNIZED;
  }
}

export function providerToJSON(object: Provider): string {
  switch (object) {
    case Provider.PROVIDER_UNSPECIFIED:
      return "PROVIDER_UNSPECIFIED";
    case Provider.STRIPE:
      return "STRIPE";
    case Provider.RAZORPAY:
      return "RAZORPAY";
    case Provider.PAYPAL:
      return "PAYPAL";
    case Provider.UNRECOGNIZED:
    default:
      return "UNRECOGNIZED";
  }
}

export enum PaymentStatus {
  UNKNOWN = 0,
  PENDING = 1,
  PAID = 2,
  FAILED = 3,
  CANCELLED = 4,
  UNRECOGNIZED = -1,
}

export function paymentStatusFromJSON(object: any): PaymentStatus {
  switch (object) {
    case 0:
    case "UNKNOWN":
      return PaymentStatus.UNKNOWN;
    case 1:
    case "PENDING":
      return PaymentStatus.PENDING;
    case 2:
    case "PAID":
      return PaymentStatus.PAID;
    case 3:
    case "FAILED":
      return PaymentStatus.FAILED;
    case 4:
    case "CANCELLED":
      return PaymentStatus.CANCELLED;
    case -1:
    case "UNRECOGNIZED":
    default:
      return PaymentStatus.UNRECOGNIZED;
  }
}

export function paymentStatusToJSON(object: PaymentStatus): string {
  switch (object) {
    case PaymentStatus.UNKNOWN:
      return "UNKNOWN";
    case PaymentStatus.PENDING:
      return "PENDING";
    case PaymentStatus.PAID:
      return "PAID";
    case PaymentStatus.FAILED:
      return "FAILED";
    case PaymentStatus.CANCELLED:
      return "CANCELLED";
    case PaymentStatus.UNRECOGNIZED:
    default:
      return "UNRECOGNIZED";
  }
}